Particle library to access the Adafruit IO REST API.
Arduino library to access the Adafruit IO REST API on the Particle
platforms like the Photon and Electron.  This is a good library for
simple send and receive access to Adafruit IO feeds.  Unlike MQTT, the
REST API does not require a connection to be kept alive so it's good for
things to periodically need to connect and send/receive values.  For
example a sensor device that sends sensor measurements to AIO and then
enters deep power-saving sleep for a period is perfect to use
with this
library.